cras: Silence pointer cast warnings for gcc -m32
This CL fixes the following warning:

> gcc -c cras_client.c -I../common -Wall -m32
> cras_client.c: In function ‘request_floop_ready’:
> cras_client.c:1827:25: warning: cast to pointer from integer of different size [-Wint-to-pointer-cast]
>  1827 |   if ((uintptr_t)req == (struct floop_request *)tag)
>       |                         ^
> cras_client.c:1827:22: warning: comparison between pointer and integer
>  1827 |   if ((uintptr_t)req == (struct floop_request *)tag)
>       |                      ^~
> cras_client.c: In function ‘request_floop’:
> cras_client.c:4194:40: warning: cast from pointer to integer of different size [-Wpointer-to-int-cast]
>  4194 |  cras_fill_request_floop(&msg, params, (uint64_t)&req);
>       |                                        ^

The tag for request_floop was used for the client to identify different
responses, if there are multiple requests in-flight concurrently.
The current implementation conveniently uses the memory address of the
request object as the tag. The tag is a uint64_t which upsets gcc
on systems with a 32-bit pointer size. An additional cast through
uintptr_t is added to silence the warning.

BUG=b:255209899
TEST=gcc -c cras_client.c -I../common -Wall -m32

Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/3970965
Tested-by: Li-Yu Yu <aaronyu@google.com>
Commit-Queue: Li-Yu Yu <aaronyu@google.com>

+++ cras/src/libcras/cras_client.c
@@ -1824,7 +1824,7 @@ static void request_floop_ready(struct cras_client *client, int32_t dev_idx,
 	pthread_mutex_lock(&client->floop_request_list_mu);
 
 	DL_FOREACH (client->floop_request_list, req) {
-		if (req == (struct floop_request *)tag)
+		if (req == (struct floop_request *)(uintptr_t)tag)
 			break;
 	}
 
@@ -4191,7 +4191,7 @@ static int32_t request_floop(struct cras_client *client,
 	// The request is tagged with the address of req which is unique
 	// The server will return the result along with the unique tag
 	struct cras_request_floop msg;
-	cras_fill_request_floop(&msg, params, (uint64_t)&req);
+	cras_fill_request_floop(&msg, params, (uint64_t)(uintptr_t)&req);
 	rc = write_message_to_server(client, &msg.header);
 	if (rc < 0) {
 		goto cleanup;
